+++ src/components/index.js
@@ -178,8 +178,6 @@ export ListingBlockBody from '@plone/volto/components/manage/Blocks/Listing/List
 export ListingBlockData from '@plone/volto/components/manage/Blocks/Listing/ListingData';
 export ListingBlockStyle from '@plone/volto/components/manage/Blocks/Listing/ListingStyle';
 export ListingBlockMore from '@plone/volto/components/manage/Blocks/Listing/ListingMore';
-export ListingBlockSidebar from '@plone/volto/components/manage/Blocks/Listing/ListingSidebar';
-export QuerystringWidget from '@plone/volto/components/manage/Blocks/Listing/QuerystringWidget';
 
 export ImageSidebar from '@plone/volto/components/manage/Blocks/Image/ImageSidebar';
 export MapsSidebar from '@plone/volto/components/manage/Blocks/Maps/MapsSidebar';